chore: Update redis package.json and add redis client configuration
This commit updates the package.json file of the redis module to version 2.8.0. It also adds a new file, 0-redis_client.js, which contains the configuration for the redis client. The purpose of this change is to ensure that the redis module is up to date and to provide a client configuration for connecting to the redis server.